fix: default story and improve the code a bit
