+++ packages/ladle/lib/cli/metro-prod.js
@@ -4,8 +4,11 @@ import importFrom from "import-from";
 
 import { entryFilePath } from "./metro-base.js";
 import { fileURLToPath } from "url";
-import metroDev, { getExtraHeaderStuff } from "./metro-dev.js";
-import { createHTMLTemplate } from "./metro/prepare-assets.js";
+import metroDev from "./metro-dev.js";
+import {
+  createHTMLTemplate,
+  getExtraHeaderStuff,
+} from "./metro/prepare-assets.js";
 import { projectPublicDir } from "./metro/utils.js";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -32,55 +35,49 @@ function prepareOutDir() {
  * @param configFolder {string}
  */
 const metroProd = async (ladleConfig, configFolder) => {
-  try {
-    // Prepare the build folder
-    const cssFile = path.resolve(appRoot, "./ladle.css");
+  // Prepare the build folder
+  const cssFile = path.resolve(appRoot, "./ladle.css");
 
-    // Make sure the out dir is cleaned and ready.
-    prepareOutDir();
+  // Make sure the out dir is cleaned and ready.
+  prepareOutDir();
 
-    const onProgress = (...args) => {
-      // console.log("Progress...", args);
-    };
+  const onProgress = (...args) => {
+    // console.log("Progress...", args);
+  };
 
-    const sourceMap = false;
+  const sourceMap = false;
 
-    const { metroServer } = await metroDev(ladleConfig, configFolder);
-    const bundle = await metroServer.build({
-      ...Server.DEFAULT_BUNDLE_OPTIONS,
-      entryFile: path.relative(projectRoot, entryFilePath),
-      dev: false,
-      minify: false,
-      platform: "web",
-      out: path.resolve(assetsDir, "ladle.js"),
-      onProgress,
+  const { metroServer } = await metroDev(ladleConfig, configFolder);
+  const bundle = await metroServer.build({
+    ...Server.DEFAULT_BUNDLE_OPTIONS,
+    entryFile: path.relative(projectRoot, entryFilePath),
+    dev: false,
+    minify: false,
+    platform: "web",
+    out: path.resolve(assetsDir, "ladle.js"),
+    onProgress,
 
-      sourceMap,
-      inlineSourceMap: false,
-      bundleType: "bundle",
-    });
+    sourceMap,
+    inlineSourceMap: false,
+    bundleType: "bundle",
+  });
 
-    const html = createHTMLTemplate({
-      appendToHead: getExtraHeaderStuff(ladleConfig, configFolder),
-      bundleUrl: "assets/ladle.js",
-      assets: [{ type: "css", filename: "assets/ladle.css" }],
-    });
+  const html = createHTMLTemplate({
+    appendToHead: getExtraHeaderStuff(ladleConfig, configFolder),
+    bundleUrl: "assets/ladle.js",
+    assets: [{ type: "css", filename: "assets/ladle.css" }],
+  });
 
-    // Manually copy/write assets
+  // Manually copy/write assets
 
-    // Copy the files under project's public dir
-    if (fs.existsSync(projectPublicDir)) {
-      fs.cpSync(projectPublicDir, outDir, { recursive: true });
-    }
-
-    fs.writeFileSync(path.resolve(assetsDir, "ladle.js"), bundle.code);
-    fs.writeFileSync(path.resolve(outDir, "index.html"), html);
-    fs.copyFileSync(cssFile, path.resolve(assetsDir, "ladle.css"));
-  } catch (e) {
-    console.log(e);
-    return false;
+  // Copy the files under project's public dir
+  if (fs.existsSync(projectPublicDir)) {
+    fs.cpSync(projectPublicDir, outDir, { recursive: true });
   }
-  return true;
+
+  fs.writeFileSync(path.resolve(assetsDir, "ladle.js"), bundle.code);
+  fs.writeFileSync(path.resolve(outDir, "index.html"), html);
+  fs.copyFileSync(cssFile, path.resolve(assetsDir, "ladle.css"));
 };
 
 export default metroProd;
