+++ packages/ladle/lib/cli/metro/prepare-assets.js
@@ -2,6 +2,24 @@ import fs from "fs";
 import path from "path";
 import { appRoot } from "./utils.js";
 
+/**
+ * @param ladleConfig {import("../../shared/types").Config}
+ * @param configFolder {string}
+ */
+export function getExtraHeaderStuff(ladleConfig, configFolder) {
+  let appendToHead = "";
+
+  const headHtmlPath = path.join(configFolder, "head.html");
+  if (fs.existsSync(headHtmlPath)) {
+    appendToHead = fs.readFileSync(headHtmlPath, "utf8");
+  }
+  if (ladleConfig.appendToHead) {
+    appendToHead += ladleConfig.appendToHead;
+  }
+
+  return appendToHead;
+}
+
 function getBaseHTMLTemplate() {
   const template = fs
     .readFileSync(path.resolve(appRoot, "index.html"))
